QLearning 보상 조건으로 에피소드 종료 테스트 및 버그 수정

+++ ColaLearning/QLearning.h
@@ -24,7 +24,7 @@ public:
 	 * \param next_StateTable : 다음 상태 표
 	 * \param reward_Table : 보상표
 	 */
-	QLearning(int reward_EndCondition, string next_StateTable, string reward_Table);
+	QLearning(double reward_EndCondition, string next_StateTable, string reward_Table);
 	/**
 	 * 모든 에피소드 초기화.
 	 *
